Version 2

+++ backend/controllers/sauces.js
@@ -1,8 +1,8 @@
 const modelSauce = require('../models/Sauce');
-const jwt = require('jsonwebtoken');
 const fs = require('fs');
-// require('dotenv').config();
 
+
+// Affiche toutes les sauces
 exports.getAllSauces = (req, res, next) => {
    modelSauce.find()
     .then(sauces => res.status(200).json(sauces))
@@ -12,32 +12,36 @@ exports.getAllSauces = (req, res, next) => {
     }));  
 };
 
+
+// Affiche une sauce selon son ID
 exports.getSauceById = (req, res, next) => {
     modelSauce.findOne({_id: req.params.id})
     .then(sauces => res.status(200).json(sauces))
     .catch(error => res.status(400).json({
         error: error, 
-        message: "Impossible de récupérer la sauce avec l'identifiant '" + req.params.id
+        message: "Impossible de récupérer la sauce avec cet identifiant"
     }));  
 };
 
+
+// Ajout d'une sauce
 exports.createSauce = (req, res, next) => {
     const sauceObject = JSON.parse(req.body.sauce);
     delete sauceObject._id;
+    delete sauceObject._userId
     const sauce = new modelSauce({
         ...sauceObject,
+        userId: req.auth.userId,
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
-        likes: 0,
-        dislikes:0,
-        usersLiked: [""],
-        usersDisliked: [""]
     });
+
     sauce.save()
     .then(() => res.status(201).json({message: "Sauce ajoutée"}))
-    .catch(error => res.status(400).json({error}));
-    console.log('Sauce initialisée');
+    .catch(error => res.status(400).json({ error }));
 };
 
+
+// Modifier une sauce
 exports.modifySauce = (req, res ,next) => {
 	const sauceObject = req.file ?{
         ...JSON.parse(req.body.sauce),
@@ -46,9 +50,10 @@ exports.modifySauce = (req, res ,next) => {
 
     modelSauce.findOne({_id: req.params.id})
         .then((sauceToModify) => {
-            if (sauceToModify.userId != req.body.userId) {
+            if (sauceToModify.userId != req.auth.userId) {
                 res.status(403).json({ message : "Vous n'avez pas le droit de modifier cette sauce"});
-            } else {
+            } 
+            else {
                 modelSauce.updateOne({ _id: req.params.id}, { ...sauceObject, _id: req.params.id})
                 .then(() => res.status(200).json({message : 'Sauce modifiée'}))
                 .catch(error => res.status(500).json({error}));
@@ -59,78 +64,56 @@ exports.modifySauce = (req, res ,next) => {
         });
 };
 
-exports.deleteSauce = (req, res, next) => { 
 
+// Supprime une sauce
+exports.deleteSauce = (req, res, next) => {
 	modelSauce.findOne({_id: req.params.id})
-    .then(sauce => {
-        const filename = sauce.imageUrl.split('/images/')[1];
-        fs.unlink(`images/${filename}`, () => {
-            modelSauce.deleteOne({_id: req.params.id})
-            .then(() => res.status(200).json({message: 'Sauce supprimée'}))
-            .catch(error => res.status(400).json({error}));  
-            console.log('Sauce supprimée'); 
-        });
+    .then((sauceToDelete) => {
+        if (sauceToDelete.userId != req.auth.userId) {
+            res.status(403).json({message : "Vous n'avez pas le droit de supprimer cette sauce"});
+        } 
+        else {
+            const filename = sauceToDelete.imageUrl.split('/images/')[1];
+            fs.unlink(`images/${filename}`, () => {
+                modelSauce.deleteOne({_id: req.params.id})
+                .then(() => res.status(200).json({message: 'Sauce supprimée'}))
+                .catch(error => res.status(400).json({ error }));  
+                console.log('Sauce supprimée'); 
+            });
+        }
     })
-    .catch(error => res.status(500).json({error}));
+    .catch(error => res.status(500).json({ error }));
 };
 
-// exports.deleteSauce = (req, res, next) => {
 
-//     const token = req.headers.authorization.split(' ')[1];
-//     const decodedToken = jwt.verify(token, process.env.RANDOM_TOKEN_SECRET);
-//     const userId = decodedToken.userId;
-//     console.log("UseridToken :", userId);
-
-// 	modelSauce.findOne({_id: req.params.id})
-//     .then((sauceToDelete) => {
-
-//         console.log("UseridSauce :", sauceToDelete.userId );
-//         console.log("CurrentUserid :", req.body.userId);
-        
-//         if (sauceToDelete.userId != userId) {
-//             res.status(403).json({message : "Vous n'avez pas le droit de supprimer cette sauce"});
-//         } else {
-//             const filename = sauce.imageUrl.split('/images/')[1];
-//             fs.unlink(`images/${filename}`, () => {
-//                 modelSauce.deleteOne({_id: req.params.id})
-//                 .then(() => res.status(200).json({message: 'Sauce supprimée'}))
-//                 .catch(error => res.status(400).json({error}));  
-//                 console.log('Sauce supprimée'); 
-//             });
-//         }
-//     })
-//     .catch(error => res.status(500).json({error}));
-// };
-
-
+// Like et dislike de la sauce
 exports.likeOrDislike = (req, res, next) => {
     let action = req.body.like;
     let sauceId = req.params.id;
-    let currentUserId = req.body.userId;
+    let currentUserId = req.auth.userId;
 
     modelSauce.findOne({ _id: sauceId })
         .then((sauceLiked) => {
             switch (action) {
                 case 1:
-                    // Vérifie si l'utilisateur n'a pas déja liké la sauce
                     if (!sauceLiked.usersLiked.includes(currentUserId)) {
                         modelSauce.updateOne(
                             {_id: sauceId}, 
                             { $push: {usersLiked: currentUserId}, $inc: {likes: +1} }
                         )
                             .then(() => res.status(200).json({message: 'Sauce likée'}))
-                            .catch(error => res.status(400).json({error}));  
+                            .catch(error => res.status(400).json({ error }));  
 
-                        // Vérifie si l'utilisateur n'a pas déja disliké la sauce
                         if (sauceLiked.usersDisliked.includes(currentUserId)){
                             modelSauce.updateOne(
                                 {_id: sauceId}, 
                                 { $pull: {usersDisliked: currentUserId}, $inc: {dislikes: -1} }
                             )
                             .then(() => res.status(200).json({message: 'Sauce dislikée'}))
-                            .catch(error => res.status(400).json({error}));  
+                            .catch(error => res.status(400).json({ error }));  
                         }
-                    }else{
+                    }
+                    else{
                         res.status(409).json({ message: 'Vous avez déja liké cette sauce' });
                     }
                     break;
@@ -142,42 +125,40 @@ exports.likeOrDislike = (req, res, next) => {
                             { $pull: {usersLiked: currentUserId}, $inc: {likes: -1} }
                         )
                             .then(() => res.status(200).json({message: 'Like retiré'}))
-                            .catch(error => res.status(400).json({error}));  
+                            .catch(error => res.status(400).json({ error }));  
                     }
                     if(sauceLiked.usersDisliked.includes(currentUserId)){
                         modelSauce.updateOne(
-                            {_id: sauceId}, 
+                            { _id: sauceId }, 
                             { $pull: {usersDisliked: currentUserId}, $inc: {dislikes: -1} }
                         )
                             .then(() => res.status(200).json({message: 'Dislike retiré'}))
-                            .catch(error => res.status(400).json({error}));  
+                            .catch(error => res.status(400).json({ error }));  
                     }
                     break;
                 
                 case -1:
-                    // Vérifie si l'utilisateur n'a pas déja disliké la sauce
                     if (!sauceLiked.usersDisliked.includes(currentUserId)) {
                         modelSauce.updateOne(
                             {_id: sauceId}, 
                             { $push: {usersDisliked: currentUserId}, $inc: {dislikes: +1} }
                         )
                             .then(() => res.status(200).json({message: 'Sauce dislikée'}))
-                            .catch(error => res.status(400).json({error}));  
+                            .catch(error => res.status(400).json({ error }));  
 
-                        // Vérifie si l'utilisateur n'a pas déja liké la sauce
                         if (sauceLiked.usersLiked.includes(currentUserId)){
                             modelSauce.updateOne(
-                                {_id: sauceId}, 
-                                { $pull: {usersLiked: currentUserId}, $inc: {likes: -1} }
+                                { _id: sauceId }, 
+                                {  $pull: {usersLiked: currentUserId}, $inc: {likes: -1} }
                             )
-                            .then(() => res.status(200).json({message: 'Like retiré'}))
-                            .catch(error => res.status(400).json({error}));  
+                            .then(() => res.status(200).json({ message: 'Like retiré' }))
+                            .catch(error => res.status(400).json({ error }));  
                         }
-                    }else{
+                    }
+                    else{
                         res.status(409).json({ message: 'Vous avez déja dislikée cette sauce' });
                     }
                     break;
             }
-            console.log("sauce", sauceLiked);
         }) 
 }
